requirements.txt
streamlit~=1.10.0
pandas~=1.0.1
tensorflow~=1.14.0
gast~=0.5.4
six~=1.16.0
termcolor~=1.1.0
numpy~=1.18.0
Keras~=2.2.5
matplotlib~=3.3.4
mpld3~=0.5.10
scikit-learn~=0.24.2
spotipy~=2.23.0
requests~=2.27.1
Pillow~=7.0.0